requirements.txt
Appium-Python-Client==2.11.1
pytest==7.4.0
allure-pytest==2.13.2
isort=5.12.0
